# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esta es la documentación para uso del API para la aplicación ramspace",
    "version": "1.0.0",
    "title": "RamSpace API",
    "contact": {
      "email": "jdany.mc@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
...
  "paths": {
...
    "/usuario": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Obtiene todas los Usuarios",
        "description": "Obtiene todas los Usuarios",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UsuarioResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Guarda un Usuario",
        "description": "Guarda un Usuario",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Usuario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Marca"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/UsuarioResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuario/registro": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Registra un Usuario",
        "description": "Registra un Usuario",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Usuario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsuarioLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/UsuarioResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Obtiene datos de un Usuario",
        "description": "Obtiene datos de un Usuario",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Usuario para mostrar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/UsuarioResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "summary": "Actualiza datos de un Usuario",
        "description": "Actualiza datos de un Usuario",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Usuario para actualizar",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Usuario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/UsuarioResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid Status Value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "summary": "Elimina un Usuario",
        "description": "Elimina un Usuario",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Usuario a eliminar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "string",
              "example": "OK"
            }
          },
          "404": {
            "description": "Invalid Status Value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usuario/login": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "summary": "Login de Usuario",
        "description": "Login de Usuario",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Marca",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsuarioLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "string",
                  "example": "token here"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Status Value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Marca": {
      "type": "object",
      "properties": {
        "descripcion": {
          "type": "string",
          "example": "Mi Marca"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "Usuario": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@mail.com"
        },
        "password": {
          "type": "string",
          "example": "admin"
        },
        "tipo": {
          "type": "integer",
          "example": 1
        },
        "primerapellido": {
          "type": "string",
          "example": "Perez"
        },
        "segundoapellido": {
          "type": "string",
          "example": "Perez"
        },
        "nombres": {
          "type": "string",
          "example": "Juan"
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "UsuarioResponse": {
      "type": "object",
      "properties": {
        "idusuario": {
          "type": "integer",
          "example": "1"
        },
        "email": {
          "type": "string",
          "example": "admin@mail.com"
        },
        "tipo": {
          "type": "integer",
          "example": 1
        },
        "primerapellido": {
          "type": "string",
          "example": "Perez"
        },
        "segundoapellido": {
          "type": "string",
          "example": "Perez"
        },
        "nombres": {
          "type": "string",
          "example": "Juan"
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "UsuarioLogin": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "admin@mail.com"
        },
        "password": {
          "type": "string",
          "example": "admin"
        }
      }
    },
...
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "example": false
        },
        "error": {
          "type": "string",
          "example": "No se encontro registro"
        }
      }
    }
  }
}